Created a generic singleton class + started framework for future games
