+++ LaserDefenderProj/Assets/Prefabs/AudioPlayer.prefab
@@ -45,7 +45,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 1fd44478d3f6d0b4ba07fbab7cca55c9, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  playerVolume: 0.4
+  playerVolume: 0.2
   enemyVolume: 0.1
   actionMusic: {fileID: 8300000, guid: c9c58b1e9787f1a45add597104d10198, type: 3}
   backgroundMusicVolume: 0.1
