+++ LaserDefenderProj/Assets/Scripts/Audio/AudioPlayer.cs
@@ -1,6 +1,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+
 public enum AudioSource_e
 {
     AUDIO_SOURCE_PLAYER,
@@ -8,7 +9,7 @@ public enum AudioSource_e
     AUDIO_SOURCE_BACKGROUND_MUSIC
 }
 
-public class AudioPlayer : MonoBehaviour
+public class AudioPlayer : DFramework.SingletonMonoBehavior<AudioPlayer>
 {
     private const float AUDIO_VOLUME_MIN = 0f;
     private const float AUDIO_VOLUME_MAX = 1f;
@@ -24,8 +25,18 @@ public class AudioPlayer : MonoBehaviour
     [SerializeField][Range(AUDIO_VOLUME_MIN, AUDIO_VOLUME_MAX)] private float backgroundMusicVolume;
 
     private AudioSource audioSourceBackground;
-    static private AudioPlayer audioPlayerSingletonInstance; // This will store reference to current object for singleton implementation
 
+    private void Start()
+    {
+        // init
+        audioSourceBackground = GetComponent<AudioSource>();
+
+        // TODO: For now, I will just have the action music playing constantly.
+        // Once I have a starting screen, I will fade between the calm->action music once the gameplay loop begins.
+        audioSourceBackground.clip = actionMusic;
+        audioSourceBackground.volume = backgroundMusicVolume;
+        audioSourceBackground.Play();
+    }
 
     public void PlaySound(AudioClip audioClip, AudioSource_e audioSource)
     {
@@ -48,39 +59,5 @@ public class AudioPlayer : MonoBehaviour
         }
 
         AudioSource.PlayClipAtPoint(audioClip, position, volume);
-    }
-
-    private void Awake()
-    {
-        ManageSingleton();
-    }
-
-    void ManageSingleton()
-    {
-        if (audioPlayerSingletonInstance == null)
-        {
-            // first instance of this game object in the main menu
-            // Save this one so that it persists on future scenes
-            audioPlayerSingletonInstance = this;
-            DontDestroyOnLoad(audioPlayerSingletonInstance);
-        }
-        else
-        {
-            // We've already createed an audio player instance, so destroy the new game object.
-            gameObject.SetActive(false);
-            Destroy(gameObject);
-        }
-    }
-
-    private void Start()
-    {
-        // init
-        audioSourceBackground = GetComponent<AudioSource>();
-
-        // TODO: For now, I will just have the action music playing constantly.
-        // Once I have a starting screen, I will fade between the calm->action music once the gameplay loop begins.
-        audioSourceBackground.clip = actionMusic;
-        audioSourceBackground.volume = backgroundMusicVolume;
-        audioSourceBackground.Play();
     }
 }
